Title A function to cope with metabolic flux

Sometimes it may be slow if the dataset is big. Just be patient. all the data used are save in /data/ file

metaflux(
  input,
  CKrepeat,
  EXPrepeat,
  EXPgroup = length(EXPrepeat),
  deltaRT = 0.01,
  ddC = 2e-04,
  Cnum = 6,
  deltaC = 1.0033
)

Arguments

input

the data you want to put in;see data/SXT-neg.csv

CKrepeat

the repeat count of Control group

EXPrepeat

the repeat count of every Control group NOTICE: it can be c(3,3) or 3. However, make sure every experiment group has the same repeat count

EXPgroup

(Optional) the number of experiment groups( except CK ); it can count automatically

deltaRT

the allowed deviation of retention time (default=0.01)

ddC

the allowed deviation of the carbon atom (default =0.0002)

Cnum

the number of carbon atoms that are marked by 13C (default =6)

deltaC

the mass of a carbon atom(default=1.0033)

Value

a data frame that shows which peak may be the result of 13C isotope labelling

Examples

# \donttest{
input<-read.csv(file = "data/20230728cell_flux_10neg.csv",header=T,check.names = F)
#> Warning: cannot open file 'data/20230728cell_flux_10neg.csv': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
neg<-metaflux(input, CKrepeat = 3, EXPrepeat = 3, Cnum = 6)
#> Error in metaflux(input, CKrepeat = 3, EXPrepeat = 3, Cnum = 6): object 'input' not found

input<-read.csv(file = "data/SXT-neg.csv",header=T,check.names = F)
#> Warning: cannot open file 'data/SXT-neg.csv': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
neg<-metaflux(input, CKrepeat = 3, EXPrepeat = c(3,3), Cnum = 6)# }
#> Error in metaflux(input, CKrepeat = 3, EXPrepeat = c(3, 3), Cnum = 6): object 'input' not found